Workaround expunge failure for SubnetRoute in session identity_map
- Issue seen was related to expunge failures for the ORM instance for
SubnetRoute in the update_subnet() workflow for host routes. The issue
is not seen consistently and the debug shows that for some reason
session_id is not populated in the ORM instance sometimes as well. In
these scenarios it is also seen that expunge() does not work on some
other object instances as well but does not trigger an exception. The
workaround overrides lazy loading of subnet related tables which fixes
the issue.
- Also return early if nothing to process to avoid an unnecessary query.

Change-Id: I973a004ca7de47bfcec441e024b164b7fca13708
(cherry picked from commit f56bc97aa7ffacfe17648aec5750b1496ea3ad7a)