[AIM] Prevent overlapping CIDRs in routed VRF
Reject adding or removing a router interface if the operation would
result in overlapping AIM Subnets within a routed AIM VRF. This can be
temporarily disabled, if necessary, by setting the
allow_routed_vrf_subnet_overlap config variable to True. Setting this
will also result in a warning being logged at startup.

A followup patch will add similar checking to the gbp-validate tool,
which should be run immediately after upgrading an existing deployment
to a new version with this feature, to make sure no overlap already
exists. If it does, it may be necessary to disable this checking until
existing overlap can be cleaned up.

Change-Id: I2de7fee0d31719293654dbe6fff032cfe1f91b9d