Make DHCP provisioning blocks conditional
Commit 3c6ecad76674e20cfa5e89ec32160acdc84f9329 added a provisioning
block for port binding, in order to ensure that the DHCP agent was
configured with the port information before it was marked as ACTIVE.
When optimized DHCP is used, there is no dependency on the DHCP agent,
as DHCP DORA traffic is handled on-hypervisor. This patch makes it so
the provisioning block is only added when optimized DHCP isn't enabled.

Change-Id: Iad74252c66a754434c5b0b181266a74bd1efebfe
(cherry picked from commit 6b73d7bd4208d49fd4df45afc3d42b107e3bd8a3)