Fix issues on sync_state display on neutron based on AIM status
The sync_state displayed on neutron resources is derived from the
corresponding AIM resources - status objects. The state is an
aggregation (merge) of the state from each of the status objs. When
resources are created - in cases where they are associated with a default
AIM resource (which is already in synced state) the state shows as being
synced on create because the other AIM resources created are yet to be
placed on sync-pending state. There are also issues with reflecting the
state accurately when the APIC is down.
- Ensure that the sync_state is kept in 'build' until the status objs
for each of the corresponding AIM resources has been processed.
- The AIM resources count is tracked based on the processing in the
extend methods without maintaining explicit mappings for each type
of resource.
- This is maintained in the resource dict for each instance so bulk
extension handling is supported.
- Add a check to condition the call to _merge_status() only if a
resource instance dict is present.
- Support for network and subnets as a first step.
- The mapping from AIM resources back to the neutron resource is
maintained as dict to go back while processing status objs. The
AIM status tracking to ensure all status objs are seen is piggybacked
on to this dict so the mapping stores a tuple (res_dict, aim_status_track)
with the DN as the key.

Change-Id: Iff81de26a6a7b777a9c4c14686edb1f242c8dd6d
(cherry picked from commit aa48985c5cf5659506ff8bf9b22bd90d039ea8dc)