[AIM] Enhance gbp-validate to detect routed subnet overlap
The gbp-validate tool will now fail validation if multiple routed
subnets that map to the same VRF overlap. Note that, unlike the
similar rejection of attempts to add or remove interfaces that result
in overlap, this is not subjet to the allow_routed_vrf_subnet_overlap
config variable, and applies to scoped as well as unscoped subnets.

Change-Id: I01346b94a491e6866bcf9581fe53028889d244ba