Disallow external subnets as router interfaces
Upstream neutron doesn't prevent subnets on external networks from
being used for router interfaces. The AIM mechanism driver treats
the boundary between unscoped internal and external networks as
a boundary for NAT. This makes the semantics confusing if subnets
on external networks are used as router interfaces and gateways,
and is therefore disallowed.

Change-Id: I164d655648c77f9a1fa12dc8927d64a5d50be356