Tempest Scenario test for Connection Tracking
The scenario will start a ping on the remote server using
floating IP in a seperate thread and start tracking the
connectivity with the remote server.

The ingress ICMP rule on the server will be deleted
and at same time then the test asserts if
connectivity breaks or not.

The testcase will pass if ping stops working after
deletion of the rule.

Change-Id: I0e45a44cf20f68136720cd508cbea28a820f8ce0