Fix ambiguity in mapping to domain in port pair workflow
While trying to get the unique domain for a port on host - we check the
HostDomainMappingV2 table for the host the port is bound to. The approach
is changed to also check the bottom bound segment for the port.

Change-Id: Ia6c7823e560c89cd41b8f499a7ee46da3f373acb
(cherry picked from commit 05c576ea746edb0f89d9bac0f83380caf4ffacac)