Fix pep8 failures seen on submitted patches
Seems like with some possible change in infra we see some
pep8 failures. These are related to Python 3 checks. Handling these
so that code is functional in either Python 2.7 or Python 3 envs.
An attempt is made to be in line with handling this in other community
projects.

Change-Id: Iec4c919bebb70365f932f23b926aa80bf4da6048
(cherry picked from commit 409c1517abe9caab8ad21877685351d4c592ad28)