Add suport for baremetal vnic_type
Bind ports of vnic_type baremetal to networks that have
opflex type static segments, using Hierarchical Port Binding
(HPB), with dynamically created VLAN type segments, as well
as networks with static or dynamic VLAN type segments.  This
includes static port configuration on EPGs or L3 Out policies
in ACI using untagged mode with native VLANs. Support for using
baremetal vnic_types with trunk ports must be addressed in a
separate patch.

Requirements for baremetal vnic_type port binding are:
1) binding:profile must have a local_link_informaiton element
2) local_link_information must contain a switch_info element
3) switch_info must be a string containing comma-
   delimitted key/value pairs, which are separated
   by a colon. Required key/value pairs are:
      o  apic_dn:<dn for static path>
      o  physical_network:<physnet for interface>
   The PhysDom for the interface can optionally be
   specified using the key/value pair:
      o  physical_domain:<name of the PhysDom for the interface>
   If the physical_domain is populated, it will be used to associate
   the named PhysDom with the EPG. If no physical_domain is provided,
   then the existing domain association behavior is used (i.e. the
   HostDomainMappingV2 entries in AIM are searched for applicable
   entries).

The local_link_information also may contain the port_id
and switch_id elements, which should specify the individual
port_id and switch_id that the baremetal VNIC connects to.
However, this information currently isn't used.

Once a VNIC type of baremetal is bound, the static path information
in the binding:profile is stored in AIM, either in an EPG or the
Interface Profile for an L3 Out policy in the case of SVI networks.

Change-Id: I43b1305de59f68d9114c5856f76a98cc72ceb18e