Support cache option for legacy GBP driver
The legacy GBP driver shares the cache code in the apic_aim
mechanism driver. This class uses an option that's only available
when the apic_aim mechanism driver is loaded. In order to ensure
backwards-compatibility for the legacy GBP driver, check to see
if the configuration options exist before trying to use it for
configuring the class.

Change-Id: I26aab66b725cea1971e8d212705d47a7890e4b42