Allow both FIP and SNAT on a single port
This should be allowed if this port has multiple IPs
either through AAP or multiple fixed_ips.

Currently once a FIP is associated to a port then it
will prevent SNAT from working anymore. This works
fine for ports having just 1 IP. However if a port is
configured with multiple IPs, then associating a FIP
to one of the IPs will break the existing SNAT
configuration for the other IPs.

Change-Id: I5bfc32a021ed6d438e227cd1c71d4701f81be296