Baremetal VNIC Trunk support
This adds support for binding baremetal VNIC trunk ports,
both the parent port and any subports that belong to the
trunk. Baremetal VNIC trunk parent ports are bound the
same way as a non-trunk baremetal VNIC port, while subports
are bound using topology information stored in the
binding:profile of the trunk parent port.

The trunk status is updated in order to reflect the health
of the trunk. Code for this was leveraged from upstream
neutron, in neutron/services/trunk/rpc/server.py. This
code wasn't used directly since it's designed to be used
with agents, and therefore is tied to RPCs. The trunk
status and subport binding code can't be invoked via RPC,
and therefore is called as part of the postcommit method
for the port_update of the parent port (i.e. during port
binding of the parent port). Since subports can be added
and removed from the trunk independent of port binding,
the mechanism driver registers for subport notification
events so that it can update the trunk status, and trigger
portbinding of the subports (note: this happens even if
the parent port isn't bound, which is basically a NO-OP).

Change-Id: I052dd3350059ecc243b9642ce15d6c989db81b52