[AIM] Convert remote_ips for SG rules properly
1. When we create a SG rule in neutron without using the
remote_group_id, by default it will set remote_ip_prefix to
None, which means to allow traffic to go through from anybody
in neutron. And this will translate to remote_ips list being
empty in AIM.

2. However when we create a SG rule with remote_group_id set,
and when this remote group hasn't been assigned to any VM
ports yet, the remote_ips list in AIM is also empty but in
this case, it means don't allow traffic from anybody to go
through.

3. To address this conflict, the solution here is to convert
remote_ips list to either 0.0.0.0/0 or ::/0 depends on ipv4 or
ipv6 when the remote_ip_prefix is None and remote_group_id is
not being used. Also need to migrate the existing SG rules
properly through a migration script.

4. Currently opflex agent is adding a flow to allow all the
traffic to go through when it sees remote_ips list being empty
for a SG rule. It will need to change its behavior to not add
that flow at all from now on when remote_ips list is empty.

Change-Id: Id5c4437117304cc76fb46b04c9261f25f0879fc2