Fix missing trunk_details for a trunk without subports
With trunk ports which have a parent port but no subports, the
trunk_details field was not set in the response for endpoint details.
Fix the db query to populate the parent_port id even if the subports
table does not have an entry for the trunk_id.

Change-Id: I9e0404abd74e590ade2c0c0adcf49678b3eed5d7
(cherry picked from commit 3624a362a10967fc5c5225d944f2e3f535ec467b)