[AIM] Add pre-existing BD to network extension
Support specifying the BridgeDomain property in the apic:distinguished_names
cisco_apic network extension. This property can only be used when creating a
neutron network, and cannot be used with SVI or external networks. The plugin
ensures that the specified BridgeDomain exists, and then takes ownership
over the Bridge Domain resource. The preexisting Bridge Domain is deleted
when the neutron network that it maps to is deleted. If the entry in the
extension database gets corrupted, the Bridge Domain cannot be repaired
correctly. A subsequent patch could add support for adding the repair of
preexisting Bridge Domain mappings using a configuration file option, if
needed.

Change-Id: I3fa1a2fd785885a43b32d8eb66c23cc409a82286