When a subnet added to a bound SVI port, ensure it is added to SVI port
When we add a subnet to a bound SVI port, we need to ensure that it is
added to the SVI port as well. When the second subnet is added to a
bound port on a SVI network, the SVI port will already be present -
either created by the user or by the plugin when the binding happens.
We check and add the subnet on the SVI port so it is present when the
static path processing happens for the bound port to create the
Interface Profile for each AF.

Similarly, when we delete a subnet from a bound port - ensure that if
this subnet is not present on the host, remove the Interface profile
for the AF.
Change-Id: I47cb516c3603eacb921527625120ed155dd2e465
(cherry picked from commit fc74d30ee86951587c8b3454af3e6afb608a3b0c)