Fix DNS Domain Name in endpoint file.
When the DNS server receives the dns domain name
it includes a period in the end -- i.e. "example.com."
This causes the client to think that the DNS is "bad"
which is what it puts in /etc/resolv.conf file.
We need to remove the period in the end of dns_domain name
before the server sends the information to the client.

Change-Id: Id56da2b3972fb1ce045577565b2eb7597dc7e257