Update subnets for SVI port corresponding to bound port
SVI ports can be created by users apriori or if it is not
present the plugin creates it.
- Cleaning up code which creates new SVI ports unnecessariy if
we have stale SVI ports present. Instead now changing to reuse
stale ports if present with an update for the subnet.
Since there are two code paths where SVI ports are updated:
1) When a port is first bound to a host - we ensure that there is
an SVI port present for that host and if not it is created with
the subnets on the net.
2) When a subnet is added to a bound port, we ensure that this subnet
is updated on all SVI ports if needed. This is now refactored to only
update the SVI port corresponding to the bound port host.

Change-Id: I356ff74acd86af7cc01f403edf98f37cad9e21cb