Fix exception with cleanup
Commit 00bf365025772d423eea7da1ddbb6211ab757aed in upstream neutron
changed the call to extend the router resource from an assignment to
an update, which means that this member must be present when creating
the router dictionary. The call to get_routers throws an exception if
specific fields are requested, which don't include the gateway. This
patch ensures that the gateway extension is present, avoiding this exception.

Change-Id: Id88eaf2fb4160d27a7a0322ddf4e4b494f3dd8c4