Fix bug where security-group rule config doesn't reflect new VMs addition
When we create security group and security group rules in
different tenants, the new VMs are not updated in the
remote IPs list for security group rules and therefore are
not reflected in HPP and OVS rules.

In this fix we make sure that the VMs are updated in the
remote IPs list even when security group and security
group rules are in different tenants.

Change-Id: Ia1fe0217fe0af766579b90ae9835ec99adaf777f