Fix DB performance with sessions
Commit 9d15b9826190c0f8018b92138e8fd4499640f98a added a method to
get objects from the DB (OVO or otherwise). This introduced a
performance problem, as seen when running "neutron port-list" with
a large number of neutron ports. On one setup with 6.5k ports, the
port-list time was on the order of several minutes. This patch makes
a minor change to this method, and imports sqlalchemy instead of just
the inspect method from that library. This reduced the same port-list
time to a low double-digit number of seconds. This follows the same
pattern that was introduced in upstream neutron (see this patch:
https://review.opendev.org/c/openstack/neutron/+/428779).

Change-Id: Id43811d2526566a466747d7e94bde8fd82430355