Fix session handling
Commit 9d15b9826190c0f8018b92138e8fd4499640f98a introduced a wrapper
method for getting sessions for objects, and changed the ML2Plus plugin
to use this new API. When adding extensions to resource dictionaries,
this API is called to get the session for the object, and if one isn't
found, a new one is retrieved from the neutron DB layer. This session
is used to start a subtransaction to complete the operation.

There are cases in neutron where the autocommit=True behavior is used
to avoid creating transactions. This is helpful when there are calls
to list resources (e.g. "get_ports"), as there are callbacks used by
the revision plugin that are invoked by the "after_commit" event in
sqlalchemy. If there is no transaction, then these callbacks aren't
invoked, which makes sense when simply listing resources.

This patch modifies the use of the wrapper method, such that it checks
to see if the session is active, and will only start a subtransaction
if the session is active. If the sesion isn't active, then the call
does the same as upstream, and relies on the autocommit=True behavior.

Change-Id: I631b31c4476ca543609ae0461feafdd27abd401f
(cherry picked from commit 6d6e01af9a49a7b486b4cdbe1f9eddc536c57943)