Remove AIM job from gate votes
Upstream support for python2 in get-pip.py was dropped, causing
the devstack gate jobs to fail (see launchpad bug 1913000). In
order to get existing patches merged (which already passed the AIM
gates), the AIM gate job is temporarily marked as non-voting. This
should be removed once the existing patches have merged, and/or when
a solution to using get-pip.py is available.

Change-Id: I5675a0e5894b8a74347f62cdb45efa89cc03a231