Add workaround for bulk port create
Commit 2dc61dfbcceaf6b85ee80188b420738476a14c36 in upstream neutron
introduced a bug for extensions. Non-bulk creation of ports passes
dictionaries to the extension drivers, but the new patch passes the
DB object instead. Fixes to this were posted upstream (see [0] and [1]),
but were never backported. This patch uses a workaround to ensure that
the objects are always dictionaries. The patch should be reverted once
backports for [0] and [1] have been created and merged.

[0]: 9c5156575dccf632195da2d25598bfce2823c3fa
[1]: 2e316157edd412d747e8533746c0974cab7dfde0

Change-Id: I2e873b7940812374ef415a2ea01a866ca0905862