Fix deletion of SVI networks
The auto-deletion of ports on SVI networks was broken in
more recent releases. The reason is that the monkey-patch
to add SVI ports to AUTO_DELETE_PORT_OWNERS wasn't taking
effect. The monkey-patch was moved to a different point
in the import process to ensure that it's applied.

Change-Id: Ib419758cc622f28aef4225d34d1d202dbef2d97a
(cherry picked from commit 1e658dbc857b77d44da54f407156f90624d872a4)