FIP Status active after dissociate.
The error happens when the FIP is dissociated from the port
and ports get deleted, which are using the VIP port's fixed
IP address as an allowed-address-pairs. The expected behavior
is that dissociation succeeds, and the final status of the
dissociated FIP is “DOWN”. Instead, they are seeing the
dissociation fail with an HTTP 404, and the final FIP status
is “ACTIVE”.
fix here is to catch and ignore "port not found" exceptions.

Change-Id: I7769371b41f390adf668f976fad9ec209b5acf69
(cherry picked from commit 7fe026d9fccd8fee66296827b83242d176275126)