Fix keystone notification listener
The notification listener for Keystone was subscribing using a pool
value other than "None". The semantics for oslo.messaging notification
listeners is that there has to be at least one listener whose pool value
is set to "None" in order to ensure that the notifications are consumed.
In order to support both environments (i.e. installations where there
are other listeners whose value is already set to "None", and
installations where there are no listeners whose value is set to
"None"), the pool value is configurable, with a default value of "None".
This ensures that the default behavior is that the notification
messages are consumed, but allows for other consumers, while still
ensuring that our notification listener will receive the messages.

Change-Id: I706ee3c4e88cb8d6ad492c1b97fe48b0392b8033