Cleanup stable branches
This patch is a vehicle for cleaning up the stable branches. The
patch to master addresses a fix that was missed when [1] was merged.
That patch was created to enable the stable/ussuri branch, but it
included a PEP8 fix which should have been a separate patch that could
have been backported through the stable branches. This patch adds the
missing fix (addresses an alias with import namespace). The backports
of this patch will include the portion of the original PEP8 fix in [1]
starting from before stable/ussuri (i.e. train through newton).

Backports of this patch will add fixes to address other issues recently
found with stable branches due to end-of-life in other projects, such
as neutron.

Drop the use of basepython = 2.7 for PEP8 jobs, as that causes the
PEP8 gate to fail.

Switch to use the upper-constraints for installing neutron, and move
upper-constraints based installations earlier in test-requirements.

Mark the openstack-tox-py27 job as non-voting. The upstream gate is
missing the pre-installation of python2.7 for stable/rocky (see [2]).
This can be reverted if the job gets fixed (although given that rocky
has already been EOL'd, that seems unlikely).

[1]: https://review.opendev.org/c/x/group-based-policy/+/752338
[2]: https://review.opendev.org/c/openstack/openstack-zuul-jobs/+/873020

Change-Id: Idfd8ccc60ed6cd0fffe63064faa3e7eb46cf8cbe