Remove explicit ep_move_detect_mode config
The AIM BridgeDomain resource was being configured with an
explicit value of the ep_move_detect_mode parameter. This was
done to address a workaround needed for older hardware. Now
that the older hardware is no longer supported, AIM has been
changed to disable this value by default, which means that the
AIM mechanism driver no longer needs to set this value explicilty.

Change-Id: I41036952c46bfd72e0c9ed2416fcf3af6294c9ad