Work around DbBasePluginCommon._get_subnet removal
The upstream commit [0] removed the _get_subnet method
from the common DB plugin. However, this method is still
in use by the group-based-policy project. As a workaround,
we re-add this method to our specialization of the core
plugin. This should be replaced at some point by a method
that follows upstream.

[0]: //review.opendev.org/c/openstack/neutron/+/742829.

Change-Id: I64615d499cca206a840a6d70477513132077fd50